The database index for jobs is now 64 bit.  If you happen to be close to 4 billion jobs in your database you will want to update your slurmctld at the same time as your slurmdbd to prevent roll over of this variable as it is 32 bit previous versions of Slurm.

+++ src/plugins/accounting_storage/mysql/as_mysql_job.h
@@ -58,7 +58,7 @@ extern int as_mysql_step_start(mysql_conn_t *mysql_conn,
 extern int as_mysql_step_complete(mysql_conn_t *mysql_conn,
 			       struct step_record *step_ptr);
 
-extern int as_mysql_suspend(mysql_conn_t *mysql_conn, uint32_t old_db_inx,
+extern int as_mysql_suspend(mysql_conn_t *mysql_conn, uint64_t old_db_inx,
 			    struct job_record *job_ptr);
 
 extern int as_mysql_flush_jobs_on_cluster(
